refactor(api-service): flatten managed-agent-setup folder
Drop the redundant `managed-agent-setup/managed-agent-setup/` nesting — leftover from the original feature commit. Files move up one level; relative imports shorten by one `../`. No behaviour change.
