+++ apps/api/src/app/agents/usecases/managed-agent-setup/managed-agent-setup-complete.command.ts
@@ -1,6 +1,6 @@
 import { BaseCommand } from '@novu/application-generic';
 
-import type { McpOAuthState } from '../../generate-mcp-oauth-url/mcp-oauth-state';
+import type { McpOAuthState } from '../generate-mcp-oauth-url/mcp-oauth-state';
 
 export class ManagedAgentSetupCompleteCommand extends BaseCommand {
   stateData: McpOAuthState;
