+++ apps/api/src/app/agents/usecases/managed-agent-setup/managed-agent-setup-inbound.command.ts
@@ -1,6 +1,6 @@
 import { IsNotEmpty, IsString } from 'class-validator';
 
-import { EnvironmentWithUserCommand } from '../../../../shared/commands/project.command';
+import { EnvironmentWithUserCommand } from '../../../shared/commands/project.command';
 
 export class ManagedAgentSetupInboundCommand extends EnvironmentWithUserCommand {
   @IsString()
