+++ apps/api/src/app/agents/usecases/managed-agent-setup/setup-card.helpers.ts
